+++ src/main/java/com/bookshop/repository/CustomerRespository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.bookshop.domain.Customer;
 
-public interface CustomerRespository extends JpaRepository<Customer, Long> {
+public interface CustomerRespository extends JpaRepository<Customer, Integer> {
 
 	Optional<Customer> findByEmail(String email);
