+++ src/main/java/com/bookshop/repository/OrderRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 
 import com.bookshop.domain.Order;
 
-public interface OrderRepository extends PagingAndSortingRepository<Order, Long> {
+public interface OrderRepository extends PagingAndSortingRepository<Order, Integer> {
 
 	Page<Order> findAllByCustomer_Id(PageRequest pageRequest, Integer id);
